requirements.txt
Flask==2.3.1
httpie==3.2.1
python-dotenv==1.0.0